Split up the evaluator. It isn't good.
